back/cdk/src/feature/audio-transcription/consumer/container/requirements.txt
demucs==4.0.0
openai-whisper==20250625
numpy
soundfile
ffmpeg-python
boto3
